+++ raiko-guest/Cargo.toml
@@ -8,6 +8,7 @@ edition = "2021"
 [dependencies]
 risc0-zkvm = { version = "0.18", default-features = false, features = ["std"] }
 zeth-lib = { path = "../lib", default-features = false, features = ["taiko"] }
+raiko-host = { path = "../raiko-host", features = ["taiko"]}
 sgx-ra = { optional = true, path = "../sgx-ra" }
 tokio = { version = "1.23", features = ["full"] }
 anyhow = "1.0"
