+++ raiko-guest/src/ratls_server.rs
@@ -1,6 +1,6 @@
 use crate::app_args::{GlobalOpts, ServerArgs};
 
-pub fn ratls_server(_: GlobalOpts, args: ServerArgs) {
+pub fn ratls_server(_: GlobalOpts, _args: ServerArgs) {
     #[cfg(feature = "sgx")]
     let _ = server_sgx::result_main(args.addr);
 }
