+++ raiko-host/src/prover/mod.rs
@@ -7,5 +7,3 @@ pub mod prepare_input;
 pub mod proof;
 pub mod request;
 pub mod server;
-pub mod taiko_extra;
-pub mod utils;
